scripts/decode_cryptedbinfile: add support for CRYPTEDB64FILE
This was only tested yet with a data sample from the past. The
hexadecimal data was converted to Base64 and the script was able
to decode this data to clear-text.

It's still unknown, whether the CRYPTEDB64FILE type uses the same
encryption scheme (AES-256 with ECB) as the earlier CRYPTEDBINFILE
type.